+++ system/helpers/cookie_helper.php
@@ -36,34 +36,6 @@ defined('BASEPATH') OR exit('No direct script access allowed');
  * @link		http://codeigniter.com/user_guide/helpers/cookie_helper.html
  */
 
-// ------------------------------------------------------------------------
-
-if ( ! function_exists('set_cookie'))
-{
-	/**
-	 * Set cookie
-	 *
-	 * Accepts seven parameters, or you can submit an associative
-	 * array in the first parameter containing all the values.
-	 *
-	 * @param	mixed
-	 * @param	string	the value of the cookie
-	 * @param	string	the number of seconds until expiration
-	 * @param	string	the cookie domain.  Usually:  .yourdomain.com
-	 * @param	string	the cookie path
-	 * @param	string	the cookie prefix
-	 * @param	bool	true makes the cookie secure
-	 * @param	bool	true makes the cookie accessible via http(s) only (no javascript)
-	 * @return	void
-	 */
-	function set_cookie($name = '', $value = '', $expire = '', $domain = '', $path = '/', $prefix = '', $secure = FALSE, $httponly = FALSE)
-	{
-		// Set the config file options
-		$CI =& get_instance();
-		$CI->input->set_cookie($name, $value, $expire, $domain, $path, $prefix, $secure, $httponly);
-	}
-}
-
 // --------------------------------------------------------------------
 
 if ( ! function_exists('get_cookie'))